// src/app/model/levels.ts
export class Levels {
    // tslint:disable-next-line: variable-name
    valueLevel_1: number;
    valueLevel_2: number;
    valueLevel_3: number;
    valueLevel_4: number;
    valueLevel_5: number;
    valueLevel_6: number;
    valueLevel_7: number;
    valueLevel_8: number;
    valueLevel_9: number;
    valueLevel_10: number;
    valueLevel_11: number;
    valueLevel_12: number;
    valueLevel_13: number;
    valueLevel_14: number;
    valueLevel_15: number;

    weightLevel_1: number;
    weightLevel_2: number;
    weightLevel_3: number;
    weightLevel_4: number;
    weightLevel_5: number;
    weightLevel_6: number;
    weightLevel_7: number;
    weightLevel_8: number;
    weightLevel_9: number;
    weightLevel_10: number;
    weightLevel_11: number;
    weightLevel_12: number;
    weightLevel_13: number;
    weightLevel_14: number;
    weightLevel_15: number;

    description_1: string;
    description_2: string;
    description_3: string;
    description_4: string;
    description_5: string;
    description_6: string;
    description_7: string;
    description_8: string;
    description_9: string;
    description_10: string;
    description_11: string;
    description_12: string;
    description_13: string;
    description_14: string;
    description_15: string;


}